Moved guestbook tests into single tests.py file
